# school-results-db/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Results & Examination API",
    "description": "Complete backend API service for Examination, Marks Management, Lifecycles, and Frozen Report Cards",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/exam-classes/{id}/submit": {
      "post": {
        "summary": "Submit an exam class (teacher, exam_controller, admin)",
        "tags": [
          "Lifecycle"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/exam-classes/{id}/verify": {
      "post": {
        "summary": "Verify exam class marks (exam_controller, principal, admin)",
        "tags": [
          "Lifecycle"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/exam-classes/{id}/send-back": {
      "post": {
        "summary": "Send exam class back to DRAFT with a reason",
        "tags": [
          "Lifecycle"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/exam-classes/{id}/publish": {
      "post": {
        "summary": "Publish exam class and generate report card snapshots",
        "tags": [
          "Lifecycle"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/exam-classes/{id}/lock": {
      "post": {
        "summary": "Lock exam class permanently (principal, admin)",
        "tags": [
          "Lifecycle"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/import/marks": {
      "post": {
        "summary": "Bulk import marks from CSV, XLSX, or JSON with temporary staging validation",
        "tags": [
          "Bulk Import"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "componentId",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "sectionId",
            "required": true
          },
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "partial",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}